refactor: 디렉터리 구조 변경 외
- 컨트롤러 이름 변경
