+++ src/main/java/com/ll/commars/domain/favorite/favorite/service/FavoriteService.java
@@ -14,9 +14,9 @@ import com.ll.commars.domain.favorite.favorite.entity.FavoriteRestaurant;
 import com.ll.commars.domain.favorite.favorite.repository.jpa.FavoriteRepository;
 import com.ll.commars.domain.favorite.favorite.repository.jpa.FavoriteRestaurantRepository;
 import com.ll.commars.domain.member.entity.Member;
-import com.ll.commars.domain.restaurant.restaurant.dto.RestaurantDto;
-import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.domain.restaurant.restaurant.repository.jpa.RestaurantRepository;
+import com.ll.commars.domain.restaurant.dto.RestaurantDto;
+import com.ll.commars.domain.restaurant.entity.Restaurant;
+import com.ll.commars.domain.restaurant.repository.jpa.RestaurantRepository;
 
 import lombok.RequiredArgsConstructor;
