--- src/main/java/com/ll/commars/domain/restaurant/businessHour/controller/ApiV1BusinessHourController.java
@@ -1,13 +1,0 @@
-package com.ll.commars.domain.restaurant.businessHour.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/restaurant/business-hour")
-public class ApiV1BusinessHourController {
-
-}
