--- src/main/java/com/ll/commars/domain/restaurant/businessHour/entity/BusinessHour.java
@@ -1,52 +1,0 @@
-package com.ll.commars.domain.restaurant.businessHour.entity;
-
-import java.time.LocalDateTime;
-
-import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.global.BaseEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "restaurant_business_hours")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class BusinessHour extends BaseEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	// 요일 ( 1: 월요일, 2: 화요일, 3: 수요일, 4: 목요일, 5: 금요일, 6: 토요일, 7: 일요일 )
-	@Column(name = "day_of_week")
-	private Integer dayOfWeek;
-
-	// 영업 시작 시간
-	@Column(name = "open_time")
-	private LocalDateTime openTime;
-
-	// 영업 종료 시간
-	@Column(name = "close_time")
-	private LocalDateTime closeTime;
-
-	// RestaurantBusinessHour와 Restaurant: 다대일
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "restaurant_id")
-	private Restaurant restaurant;
-}
