--- src/main/java/com/ll/commars/domain/restaurant/businessHour/repository/jpa/BusinessHourRepository.java
@@ -1,9 +1,0 @@
-package com.ll.commars.domain.restaurant.businessHour.repository.jpa;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ll.commars.domain.restaurant.businessHour.entity.BusinessHour;
-
-public interface BusinessHourRepository extends JpaRepository<BusinessHour, Long> {
-
-}
