--- src/main/java/com/ll/commars/domain/restaurant/category/repository/jpa/RestaurantCategoryRepository.java
@@ -1,9 +1,0 @@
-package com.ll.commars.domain.restaurant.category.repository.jpa;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.ll.commars.domain.restaurant.category.entity.RestaurantCategory;
-
-public interface RestaurantCategoryRepository extends JpaRepository<RestaurantCategory, Long> {
-
-}
