+++ src/main/java/com/ll/commars/domain/restaurant/controller/BusinessHourController.java
@@ -1,0 +1,13 @@
+package com.ll.commars.domain.restaurant.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/restaurant/business-hour")
+public class BusinessHourController {
+
+}
