+++ src/main/java/com/ll/commars/domain/restaurant/entity/BusinessHour.java
@@ -1,7 +1,10 @@
-package com.ll.commars.domain.favorite.favorite.entity;
+package com.ll.commars.domain.restaurant.entity;
 
-import com.ll.commars.domain.restaurant.entity.Restaurant;
+import java.time.LocalDateTime;
 
+import com.ll.commars.global.BaseEntity;
+
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -17,24 +20,31 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@Table(name = "favorites_restaurants")
+@Table(name = "restaurant_business_hours")
 @Getter
 @Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class FavoriteRestaurant {
+public class BusinessHour extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	// FavoriteRestaurant와 Favorite: 다대일
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "favorite_id")
-	private Favorite favorite;
+	// 요일 ( 1: 월요일, 2: 화요일, 3: 수요일, 4: 목요일, 5: 금요일, 6: 토요일, 7: 일요일 )
+	@Column(name = "day_of_week")
+	private Integer dayOfWeek;
 
-	// FavoriteRestaurant와 Restaurant: 다대일
+	// 영업 시작 시간
+	@Column(name = "open_time")
+	private LocalDateTime openTime;
+
+	// 영업 종료 시간
+	@Column(name = "close_time")
+	private LocalDateTime closeTime;
+
+	// RestaurantBusinessHour와 Restaurant: 다대일
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "restaurant_id")
 	private Restaurant restaurant;
