+++ src/main/java/com/ll/commars/domain/restaurant/entity/RestaurantDoc.java
@@ -1,0 +1,52 @@
+package com.ll.commars.domain.restaurant.entity;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.elasticsearch.annotations.Document;
+import org.springframework.data.elasticsearch.annotations.Field;
+import org.springframework.data.elasticsearch.annotations.FieldType;
+import org.springframework.data.elasticsearch.annotations.GeoPointField;
+import org.springframework.data.elasticsearch.annotations.Mapping;
+import org.springframework.data.elasticsearch.annotations.Setting;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Document(indexName = "es_restaurants", createIndex = true)
+@Setting(settingPath = "elasticsearch/settings.json")
+@Mapping(mappingPath = "elasticsearch/mappings.json")
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class RestaurantDoc {
+
+	@Id
+	private String id;
+
+	@Field(type = FieldType.Text)
+	private String name;
+
+	@Field(type = FieldType.Text)
+	private String details;
+
+	@JsonProperty("average_rate")
+	@Field(name = "average_rate", type = FieldType.Double)
+	private Double averageRate;
+
+	@GeoPointField
+	private String location;
+
+	@Field(type = FieldType.Double)
+	private Double lat;
+
+	@Field(type = FieldType.Double)
+	private Double lon;
+}
