+++ src/main/java/com/ll/commars/domain/restaurant/entity/RestaurantMenu.java
@@ -1,7 +1,8 @@
-package com.ll.commars.domain.favorite.favorite.entity;
+package com.ll.commars.domain.restaurant.entity;
 
-import com.ll.commars.domain.restaurant.entity.Restaurant;
+import com.ll.commars.global.BaseEntity;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -10,6 +11,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,24 +19,32 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Entity
-@Table(name = "favorites_restaurants")
+@Table(name = "restaurants_menus")
 @Getter
 @Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class FavoriteRestaurant {
+public class RestaurantMenu extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	// FavoriteRestaurant와 Favorite: 다대일
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "favorite_id")
-	private Favorite favorite;
+	// 메뉴 이름
+	@NotNull
+	@Column(name = "name", nullable = false)
+	private String name;
 
-	// FavoriteRestaurant와 Restaurant: 다대일
+	@NotNull
+	@Column(name = "price", nullable = false)
+	private Integer price;
+
+	// 메뉴 사진
+	@Column(name = "image_url")
+	private String imageUrl;
+
+	// RestaurantMenu와 Restaurant: 다대일
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "restaurant_id")
 	private Restaurant restaurant;
