--- src/main/java/com/ll/commars/domain/restaurant/menu/entity/RestaurantMenu.java
@@ -1,52 +1,0 @@
-package com.ll.commars.domain.restaurant.menu.entity;
-
-import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.global.BaseEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "restaurants_menus")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class RestaurantMenu extends BaseEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	// 메뉴 이름
-	@NotNull
-	@Column(name = "name", nullable = false)
-	private String name;
-
-	@NotNull
-	@Column(name = "price", nullable = false)
-	private Integer price;
-
-	// 메뉴 사진
-	@Column(name = "image_url")
-	private String imageUrl;
-
-	// RestaurantMenu와 Restaurant: 다대일
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "restaurant_id")
-	private Restaurant restaurant;
-}
