+++ src/main/java/com/ll/commars/domain/restaurant/service/BusinessHourService.java
@@ -1,0 +1,20 @@
+package com.ll.commars.domain.restaurant.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.ll.commars.domain.restaurant.repository.jpa.BusinessHourRepository;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class BusinessHourService {
+
+	private final BusinessHourRepository businessHourRepository;
+
+	@Transactional
+	public void truncate() {
+		businessHourRepository.deleteAll();
+	}
+}
