+++ src/main/java/com/ll/commars/domain/review/service/ReviewService.java
@@ -9,9 +9,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.ll.commars.domain.member.entity.Member;
 import com.ll.commars.domain.member.repository.jpa.MemberRepository;
 import com.ll.commars.domain.member.service.MemberService;
-import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.domain.restaurant.restaurant.repository.jpa.RestaurantRepository;
-import com.ll.commars.domain.restaurant.restaurant.service.RestaurantService;
+import com.ll.commars.domain.restaurant.entity.Restaurant;
+import com.ll.commars.domain.restaurant.repository.jpa.RestaurantRepository;
+import com.ll.commars.domain.restaurant.service.RestaurantService;
 import com.ll.commars.domain.review.dto.ReviewDto;
 import com.ll.commars.domain.review.entity.Review;
 import com.ll.commars.domain.review.repository.jpa.ReviewRepository;
