+++ src/main/java/com/ll/commars/domain/todayRandom/controller/ApiV1TodayRandomController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ll.commars.domain.restaurant.restaurant.dto.RestaurantSummaryDTO;
+import com.ll.commars.domain.restaurant.dto.RestaurantSummaryDTO;
 import com.ll.commars.domain.review.dto.RestaurantReviewAnalysisDTO;
 import com.ll.commars.domain.todayRandom.service.TodayRandomService;
