+++ src/main/java/com/ll/commars/domain/todayRandom/service/TodayRandomService.java
@@ -14,9 +14,9 @@ import com.ll.commars.domain.favorite.favorite.entity.Favorite;
 import com.ll.commars.domain.favorite.favorite.entity.FavoriteRestaurant;
 import com.ll.commars.domain.favorite.favorite.repository.jpa.FavoriteRepository;
 import com.ll.commars.domain.favorite.favorite.repository.jpa.FavoriteRestaurantRepository;
-import com.ll.commars.domain.restaurant.restaurant.dto.RestaurantSummaryDTO;
-import com.ll.commars.domain.restaurant.restaurant.entity.Restaurant;
-import com.ll.commars.domain.restaurant.restaurant.repository.jpa.RestaurantRepository;
+import com.ll.commars.domain.restaurant.dto.RestaurantSummaryDTO;
+import com.ll.commars.domain.restaurant.entity.Restaurant;
+import com.ll.commars.domain.restaurant.repository.jpa.RestaurantRepository;
 import com.ll.commars.domain.review.dto.RestaurantReviewAnalysisDTO;
 import com.ll.commars.domain.review.dto.ReviewAnalysisDTO;
 import com.ll.commars.domain.review.entity.Review;
